OCT-111: moved GetLocalesQuery
