+++ components/catalogs/back/src/Application/Persistence/Locale/GetLocalesQueryInterface.php
@@ -2,18 +2,16 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Catalogs\Application\Persistence;
+namespace Akeneo\Catalogs\Application\Persistence\Locale;
 
 /**
  * @copyright 2022 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-interface GetLocalesByCodeQueryInterface
+interface GetLocalesQueryInterface
 {
     /**
-     * @param array<string> $codes
-     *
      * @return array<array-key, array{code: string, label: string}>
      */
-    public function execute(array $codes, int $page = 1, int $limit = 20): array;
+    public function execute(int $page = 1, int $limit = 20): array;
 }
