+++ components/catalogs/back/tests/Integration/Infrastructure/Persistence/GetLocalesByCodeQueryTest.php
@@ -4,14 +4,14 @@ declare(strict_types=1);
 
 namespace Akeneo\Catalogs\Test\Integration\Infrastructure\Persistence;
 
-use Akeneo\Catalogs\Infrastructure\Persistence\GetLocalesByCodeQuery;
+use Akeneo\Catalogs\Infrastructure\Persistence\Locale\GetLocalesByCodeQuery;
 use Akeneo\Catalogs\Test\Integration\IntegrationTestCase;
 
 /**
  * @copyright 2022 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  *
- * @covers \Akeneo\Catalogs\Infrastructure\Persistence\GetLocalesByCodeQuery
+ * @covers \Akeneo\Catalogs\Infrastructure\Persistence\Locale\GetLocalesByCodeQuery
  */
 class GetLocalesByCodeQueryTest extends IntegrationTestCase
 {
